Refactor app into multiple files
